fix(a11y): safe areas, contrast AA, colors distingibles i àrees de toc
La inserció superior de l'iPhone estava a #root, i allà s'equivocava de
tres maneres alhora. Una capçalera sticky s'enganxava a la y=0 de la
finestra, que en un iPhone amb osca és darrere del rellotge. Cada
min-h-dvh vivia dins d'una caixa amb padding, o sigui que el document
feia 100dvh + 59 px i cada pantalla curta tenia 59 px de scroll mort. I
la càmera de l'escàner, que ha de sagnar fins a dalt, començava 59 px
avall amb una franja de fons de l'app a sobre i el botó d'alta manual 59
px per sota del plec. Ara la inserció la posa cada pantalla al seu
element de dalt: una capçalera sticky pinta A TRAVÉS de la barra d'estat
i hi deixa el contingut a sota, i la càmera arriba al vidre.

Els colors, tots mesurats i no estimats: un script converteix els oklch a
sRGB i calcula els contrastos WCAG.

Set parells de text no arribaven a 4.5:1. Els fons de marca que porten
text passen a --ds-brand-cta (3.92 → 4.63), --ds-brand-hover deixa de ser
--ds-brand-strong (3.17 → 5.26) i --ds-text-faint puja a 4.58 sobre la
superfície on es fa servir.

Dos parells eren indistingibles amb brillo baix i cap dels dos ho era per
la tonalitat: --ds-warning-deep tenia EXACTAMENT la mateixa lluminositat
que --ds-success (1.000:1 en escala de grisos) i són les files pagat i
pendent de la mateixa llista; --ds-error quedava a 1.13:1 de
--ds-brand-accent, que és la cella que hi ha just sobre la targeta del
veredicte a l'escàner. La prova ara ho vigila per lluminositat i no només
per tonalitat, i compara angles com el que són, un cercle.

I el doble bip volia dir dues coses contràries: 'ja ha entrat' i 'ja
havia entrat i no torna a entrar' vibraven igual.

Àrees de toc: quatre controls de la porta es quedaven just al mínim quan
allà l'han de superar, i dotze més no hi arribaven —les pastilles del
rànquing feien 30 px, els botons d'idioma 34 i sis enllaços no tenien cap
alçada mínima.

+++ src/design/states.ts
@@ -89,7 +89,10 @@ export const SCAN_PRESENTATION = {
   },
   already_checked_in: {
     icon: 'cross',
-    haptic: HAPTIC.double,
+    // Long, not double. `ok_walkin_review` is also warning-toned and was also
+    // a double buzz, so the one signal meant to work without looking said the
+    // same thing for "they are in" and "they are not".
+    haptic: HAPTIC.long,
     tone: 'warning',
     admitted: false,
     messageKey: 'scanner.alreadyCheckedIn',
